pharmacyClient/requirements.txt
lxml==4.6.2
PyMySQL==1.0.2
pdfkit==0.6.1
pandas==1.2.1
Pillow==8.1.1
suds_jurko==0.6
